Add account model overriding support
* Create Concerns::Models::Account module
* http://guides.rubyonrails.org/engines.html
  #overriding-models-and-controllers
