+++ lib/locker_room.rb
@@ -1,6 +1,7 @@
 require "locker_room/engine"
 
 require "locker_room/concerns/scoped_to"
+require "locker_room/concerns/models/account"
 require "locker_room/concerns/models/user"
 require "locker_room/concerns/models/member"
